Security revision. The socket connection in the network connector is now a privileged action. Permission checks are now performed when creating certain key networking objects. SDBTrustedObject interface removed.
The intention of this change is to make security less confusing. Now, instead of
global scope privileged objects, the necessary security checks are performed at
object construction after which it is assumed the object is kept within a secure
context. Note that permissions checks are rather crude. A security manager is
only able to prevent MckoiDDBClient, NetworkConfigResource, NetworkProfile and
DataAddress construction. A permission check is also now on the 'all network
paths' query.

In the future there may be additional permission checks on data model
session and transaction construction and commits (writes).